:art: move vsl parsing code into their respective files.
